feat: :art: Formatted code with ESLint
Fixed any errors and format issues.
